Changing the EventHandler names to be NotificationHandler names. Moving custom hooks to a better organized folder structure.
