+++ client/src/api/api-gateway.ts
@@ -1,6 +1,6 @@
-import { ErrorHandler } from './error-handler.js';
+import { NotificationHandler } from './notification-handler.js';
 import BASE_URL from '../config/api-root.js';
-import EventHandlerTypes from '../config/event-handler-types.js';
+import NotificationHandlerTypes from '../config/notification-handler-types.js';
 
 function ApiGateway() {
     async function request(endpoint, options = {}) {
@@ -35,9 +35,9 @@ function ApiGateway() {
             // Return parsed JSON data
             return await response.json();
         } catch (error) {
-            const status = EventHandlerTypes.ERROR;
+            const status = NotificationHandlerTypes.ERROR;
             const message = error.message || 'API Gateway Error';
-            ErrorHandler(status, message);
+            NotificationHandler(status, message);
             throw error; // Re-throw error for caller to handle
         }
     }
@@ -53,9 +53,9 @@ function ApiGateway() {
             ...options,
         });
     }
-    
+
     function put(endpoint, body, options = {}) {
-        console.log('2',body);
+        console.log('2', body);
         return request(endpoint, {
             method: 'PUT',
             body: JSON.stringify(body),
