--- client/src/config/event-handler-types.js
@@ -1,6 +1,0 @@
-const EventHandlerTypes = {
-    ERROR: 'error',
-    SUCSSESS: 'success'
-};
-
-export default EventHandlerTypes;
