+++ client/src/config/notification-handler-types.js
@@ -1,0 +1,6 @@
+const NotificationHandlerTypes = {
+    ERROR: 'error',
+    SUCSSESS: 'success',
+};
+
+export default NotificationHandlerTypes;
