+++ client/src/store/slices/book.ts
@@ -2,8 +2,8 @@ import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { BookDTO } from '../../domains/books/core/dtos/book.dto';
 import BookService from '../../domains/books/index';
 import StorePaths from '../../domains/books/core/constants/book-store-paths.constants';
-import { ErrorHandler } from '../../api/error-handler';
-import EventHandlerTypes from '../../config/event-handler-types';
+import { NotificationHandler } from '../../api/notification-handler';
+import NotificationHandlerTypes from '../../config/notification-handler-types';
 
 const initialState = {
     data: <BookDTO | {}>{},
@@ -57,8 +57,10 @@ const book = createSlice({
                 state.data = action.payload;
             })
             .addCase(GET_BOOK.rejected, (state, action) => {
-                console.log(state, action);
-                ErrorHandler(EventHandlerTypes.ERROR, 'Error getting Book');
+                NotificationHandler(
+                    NotificationHandlerTypes.ERROR,
+                    'Error getting Book',
+                );
             });
 
         builder
@@ -66,11 +68,16 @@ const book = createSlice({
                 console.log('calling');
             })
             .addCase(CREATE_BOOK.fulfilled, (state, action) => {
-                console.log('Book created');
+                NotificationHandler(
+                    NotificationHandlerTypes.SUCSSESS,
+                    'Book created',
+                );
             })
             .addCase(CREATE_BOOK.rejected, (state, action) => {
-                console.log(state, action);
-                ErrorHandler(EventHandlerTypes.ERROR, 'Error getting Book');
+                NotificationHandler(
+                    NotificationHandlerTypes.ERROR,
+                    'Error getting Book',
+                );
             });
 
         builder
@@ -78,11 +85,16 @@ const book = createSlice({
                 console.log('calling');
             })
             .addCase(UPDATE_BOOK.fulfilled, () => {
-                console.log('Book updated');
+                NotificationHandler(
+                    NotificationHandlerTypes.SUCSSESS,
+                    'Book updated',
+                );
             })
             .addCase(UPDATE_BOOK.rejected, (state, action) => {
-                console.log(state, action);
-                ErrorHandler(EventHandlerTypes.ERROR, 'Error updating Book');
+                NotificationHandler(
+                    NotificationHandlerTypes.ERROR,
+                    'Error updating Book',
+                );
             });
 
         builder
@@ -90,11 +102,16 @@ const book = createSlice({
                 console.log('calling');
             })
             .addCase(DELETE_BOOK.fulfilled, () => {
-                console.log('Book deleted');
+                NotificationHandler(
+                    NotificationHandlerTypes.SUCSSESS,
+                    'Book deleted',
+                );
             })
             .addCase(DELETE_BOOK.rejected, (state, action) => {
-                console.log(state, action);
-                ErrorHandler(EventHandlerTypes.ERROR, 'Error deleting Book');
+                NotificationHandler(
+                    NotificationHandlerTypes.ERROR,
+                    'Error deleting Book',
+                );
             });
     },
 });
